rename Transport class into ZigateTransport
