+++ Classes/ZigateTransport/instrumentation.py
@@ -1,4 +1,3 @@
-import Domoticz
 import time
 
 instrument_time = True
